fix(fee): sync uint size for event

+++ contracts/interfaces/IFeeController.sol
@@ -20,7 +20,7 @@ interface IFeeController {
 
     // ================ Events =================
 
-    event SetLendingFee(bytes32 indexed id, uint64 fee);
+    event SetLendingFee(bytes32 indexed id, uint16 fee);
 
     event SetVaultMintFee(uint64 fee);
 
@@ -29,7 +29,7 @@ interface IFeeController {
     function setLendingFee(bytes32 id, uint16 fee) external;
 
     function setVaultMintFee(uint64 fee) external;
-    
+
     function getLendingFee(bytes32 id) external view returns (uint16);
 
     function getVaultMintFee() external view returns (uint64);
